+++ mscore/pianoroll/pianoroll.cmake
@@ -1,0 +1,18 @@
+
+
+set(PIANOROLL_SRC
+    ${CMAKE_CURRENT_LIST_DIR}/pianoroll.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pianoroll.h
+    ${CMAKE_CURRENT_LIST_DIR}/pianokeyboard.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pianokeyboard.h
+    ${CMAKE_CURRENT_LIST_DIR}/pianolevels.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pianolevels.h
+    ${CMAKE_CURRENT_LIST_DIR}/pianolevelschooser.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pianolevelschooser.h
+    ${CMAKE_CURRENT_LIST_DIR}/pianolevelsfilter.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pianolevelsfilter.h
+    ${CMAKE_CURRENT_LIST_DIR}/pianoruler.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pianoruler.h
+    ${CMAKE_CURRENT_LIST_DIR}/pianoview.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pianoview.h
+    )
