+++ mscore/pianoroll/pianoview.cpp
